+++ spec/gocoin/api/merchants/payouts_spec.rb
@@ -44,14 +44,6 @@ describe Gocoin::Merchants::Payouts do
 
 	end
 
-	describe "'request' method" do
-
-    it 'should return the correct result' do
-    	@payouts.request(@merchant_id, @request_params).should == @request_api_return_hash
-    end
-
-	end
-
 	describe "'list' method" do
 
     it 'should return the correct result' do